requirements.txt
# Core dependencies
psycopg2-binary>=2.9.0
boto3>=1.34.0
botocore>=1.34.0
requests>=2.28.0
python-dotenv>=1.0.0

# Web crawling
crawl4ai>=0.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Async support
aiohttp>=3.8.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# API server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Caching
redis>=4.5.0

# Utilities
python-multipart>=0.0.6
httpx>=0.24.0

# Git operations
GitPython>=3.1.0
dulwich>=0.21.0

# Code analysis
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0

# MCP Server
mcp>=0.1.0

# Development/Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0